test: move relationship order tests into OrderAndPaginationExporterTest
some time back we moved all order and pagination related tests into this test class.
