+++ dhis-2/dhis-test-integration/src/test/java/org/hisp/dhis/tracker/export/relationship/RelationshipServiceTest.java
@@ -31,17 +31,14 @@ import static org.hisp.dhis.tracker.TrackerType.ENROLLMENT;
 import static org.hisp.dhis.tracker.TrackerType.EVENT;
 import static org.hisp.dhis.tracker.TrackerType.TRACKED_ENTITY;
 import static org.hisp.dhis.utils.Assertions.assertContainsOnly;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
-import org.hisp.dhis.common.BaseIdentifiableObject;
 import org.hisp.dhis.common.CodeGenerator;
 import org.hisp.dhis.common.IdentifiableObjectManager;
-import org.hisp.dhis.common.SortDirection;
 import org.hisp.dhis.commons.util.RelationshipUtils;
 import org.hisp.dhis.feedback.ForbiddenException;
 import org.hisp.dhis.feedback.NotFoundException;
@@ -62,7 +59,6 @@ import org.hisp.dhis.trackedentity.TrackedEntity;
 import org.hisp.dhis.trackedentity.TrackedEntityType;
 import org.hisp.dhis.user.User;
 import org.hisp.dhis.user.UserService;
-import org.joda.time.DateTime;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -275,56 +271,12 @@ class RelationshipServiceTest extends SingleSetupIntegrationTestBase {
         relationships.stream().map(Relationship::getUid).collect(Collectors.toList()));
   }
 
-  @Test
-  void shouldOrderRelationshipsByUpdatedAtClientInDescOrder()
-      throws ForbiddenException, NotFoundException {
-    Relationship relationshipA = relationship(teA, teB, DateTime.now().toDate());
-    Relationship relationshipB = relationship(teA, eventA, DateTime.now().minusDays(1).toDate());
-
-    RelationshipOperationParams operationParams =
-        RelationshipOperationParams.builder()
-            .type(TRACKED_ENTITY)
-            .identifier(teA.getUid())
-            .orderBy("createdAtClient", SortDirection.DESC)
-            .build();
-    List<String> relationshipIds =
-        relationshipService.getRelationships(operationParams).stream()
-            .map(BaseIdentifiableObject::getUid)
-            .collect(Collectors.toList());
-
-    assertEquals(List.of(relationshipA.getUid(), relationshipB.getUid()), relationshipIds);
-  }
-
-  @Test
-  void shouldOrderRelationshipsByUpdatedAtClientInAscOrder()
-      throws ForbiddenException, NotFoundException {
-    Relationship relationshipA = relationship(teA, teB, DateTime.now().toDate());
-    Relationship relationshipB = relationship(teA, eventA, DateTime.now().minusDays(1).toDate());
-
-    RelationshipOperationParams operationParams =
-        RelationshipOperationParams.builder()
-            .type(TRACKED_ENTITY)
-            .identifier(teA.getUid())
-            .orderBy("createdAtClient", SortDirection.ASC)
-            .build();
-    List<String> relationshipIds =
-        relationshipService.getRelationships(operationParams).stream()
-            .map(BaseIdentifiableObject::getUid)
-            .collect(Collectors.toList());
-
-    assertEquals(List.of(relationshipB.getUid(), relationshipA.getUid()), relationshipIds);
-  }
-
   private Relationship relationship(TrackedEntity from, TrackedEntity to) {
     return relationship(from, to, teToTeType, new Date());
   }
 
-  private Relationship relationship(TrackedEntity from, TrackedEntity to, Date createdAtClient) {
-    return relationship(from, to, teToTeType, createdAtClient);
-  }
-
-  private Relationship relationship(TrackedEntity from, TrackedEntity to, RelationshipType type) {
-    return relationship(from, to, type, new Date());
+  private void relationship(TrackedEntity from, TrackedEntity to, RelationshipType type) {
+    relationship(from, to, type, new Date());
   }
 
   private Relationship relationship(
@@ -358,10 +310,6 @@ class RelationshipServiceTest extends SingleSetupIntegrationTestBase {
     manager.save(relationship);
 
     return relationship;
-  }
-
-  private Relationship relationship(TrackedEntity from, Event to, Date createdAtClient) {
-    return relationship(from, to, teToEvType, createdAtClient);
   }
 
   private Relationship relationship(TrackedEntity from, Event to) {
